refactor: FolderPage.css 생성하여 FolderPage에 적용
