+++ src/App.js
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import SharedPage from "./pages/SharedPage";
-import FolderPage from "./pages/FolderPage";
+import FolderPage from "./pages/FolderPage/FolderPage";
 import "./styles/global.css";
 
 const App = () => {
